fix!: pass arguments to non-main modules as array
Arguments are passed to modules as a string array generally. And for
main modules, they are originally taken from the command line of the
dutctl command (passed via RPC to the dutagent and finally to the
module). From the earliest occurrence the arguments are stored as an
array of strings as the go runtime passes the command line arguments (
of dutctl in this case) to the program in such way.

Now for non-main modules, arguments can be passed to modules via the
dutagent config file. For convenience, the respective config field was
designed to be a string. As a result, this string has to be split up
into a list of arguments. However, just splitting the string based on
spaces is not sufficient because this does not consider e.g. quoting.

An elegant solution would introduce a lexer that splits up the string
the same way the command line ends up in os.Args array. But for now,
this fix just changes the respective field in the dutagent config into
a string array.

BREAKING CHANGE: `args` field of module object in dutagent config is
now a string array.

Signed-off-by: Jens Drenhaus <jens.drenhaus@blindspot.software>