refactor: improve GPIO module backend abstraction
With the current abstraction the backend implementations of the GPIO
module include business logic. The backend abstracted the button and
switch, not the GPIO operation. Thus, the code of the module itself are
just shallow pass-through modules and new backends would need to
implement the business logic of button and switch again.

This refactoring lifts the business logic into the actual functions of
the GPIO modules and changes the interface of the backend to only
abstract the pure GPIO operation.

Signed-off-by: Jens Drenhaus <jens.drenhaus@blindspot.software>