ci: build and test per commit in PRs
To ensure a consistent and stable tree, the CI pipeline builds
and tests every commit in a PR. For pushes to main only the
final commit is considered for performance reasons.

The requirement that every commit in the tree builds prepares for
easy use of e.g. git bisect when finding bugs.

Further, the tests for different architectures are also integrated
in the introduced matrix strategy.

Signed-off-by: Jens Drenhaus <jens.drenhaus@blindspot.software>