fix!: remove extra newline for print messages
The client code used to add an extra newline character (via fmt.Fprintln)
to RunResponse_Print messages. This commit removes that extra newline
and outputs the received string as is. Formatting should happen in
the module that issues the Print message.

BREAKING CHANGE: The client does not add an extra newline to received
Print messages anymore.

Signed-off-by: Jens Drenhaus <jens.drenhaus@blindspot.software>