docs: fix typos, spelling and tables
This patch focuses on fixing spelling and typos in the READMEs. It also caps the line length at
around 120 characters. This was observed in some files and allows better readability in text form.

In some cases DUT Control was renamed to DUT Client where the text did not reference the whole
project but the command line tool.

Signed-off-by: Christian Grönke <christian.groenke@9elements.com>