refactor: use dutagent.Stream as project wide RPC stream abstraction
The dutagent.Stream interface was already used in the broker as an
abstranction to the connect.BidiStram. It turns out this could be
useful at other places in the project, too.

If more functionality of the connect package implementation is needed,
like header manipulation, further interfaces can be composed.

The dutagend.Stream interface focuses on message flow only and helps
parts of the code that deals with message flow logic to be cleaner and
most of all, testable.

Signed-off-by: Jens Topp <jens.topp@blindspot.software>